+++ packages/components/src/Combobox/Combobox.types.ts
@@ -143,8 +143,10 @@ export interface ComboboxOptionProps {
    * The function receives the option's props, and a boolean indicating if the option is selected.
    */
   readonly customRender?: (
-    option: Omit<ComboboxOptionProps, "customRender"> & {
+    option: Pick<ComboboxOptionProps, "id" | "label" | "prefix"> & {
       isSelected: boolean;
+      /** Render the default option content. */
+      defaultContent: ReactElement;
     },
   ) => React.ReactNode;