+++ packages/components/src/Menu/Menu.tsx
@@ -2,8 +2,12 @@ import type { CSSProperties, MouseEvent, ReactElement, RefObject } from "react";
 import React, { useId, useRef, useState } from "react";
 import classnames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
-import { useRefocusOnActivator } from "@jobber/hooks/useRefocusOnActivator";
-import { useWindowDimensions } from "@jobber/hooks/useWindowDimensions";
+import {
+  useFocusTrap,
+  useIsMounted,
+  useRefocusOnActivator,
+  useWindowDimensions,
+} from "@jobber/hooks";
 import type { IconColorNames, IconNames } from "@jobber/design";
 import {
   FloatingPortal,
@@ -15,8 +19,6 @@ import {
   useFloating,
   useInteractions,
 } from "@floating-ui/react";
-import { useFocusTrap } from "@jobber/hooks/useFocusTrap";
-import { useIsMounted } from "@jobber/hooks/useIsMounted";
 import styles from "./Menu.module.css";
 import { Button } from "../Button";
 import { Typography } from "../Typography";