+++ packages/site/src/content/AutoLink/index.tsx
@@ -1,4 +1,4 @@
-import Content from "@atlantis/docs/components/AutoLink/AutoLink.stories.mdx";
+import Content from "./AutoLink.stories.mdx";
 import MobileProps from "./AutoLink.props-mobile.json";
 import Notes from "./AutoLinkNotes.mdx";
 import { ContentExport } from "../../types/content";