+++ packages/site/src/content/Chips/index.tsx
@@ -1,4 +1,4 @@
-import ChipsContent from "@atlantis/docs/components/Chips/Chips.stories.mdx";
+import ChipsContent from "./Chips.stories.mdx";
 import Props from "./Chips.props.json";
 import Notes from "./ChipsNotes.mdx";
 import { ContentExport } from "../../types/content";