+++ packages/site/src/content/Combobox/index.tsx
@@ -1,4 +1,4 @@
-import Content from "@atlantis/docs/components/Combobox/Combobox.stories.mdx";
+import Content from "./Combobox.stories.mdx";
 import Props from "./Combobox.props.json";
 import Notes from "./ComboboxNotes.mdx";
 import { ContentExport } from "../../types/content";