+++ packages/site/src/content/ContentOverlay/index.tsx
@@ -1,4 +1,4 @@
-import Content from "@atlantis/docs/components/ContentOverlay/ContentOverlay.stories.mdx";
+import Content from "./ContentOverlay.stories.mdx";
 import MobileProps from "./ContentOverlay.props-mobile.json";
 import Notes from "./ContentOverlayNotes.mdx";
 import { ContentExport } from "../../types/content";